refactor: split validation.ts into domain modules (#312)
Split apps/backend/src/lib/validation.ts (1513 lines, 23 importers)
into 16 domain modules under validation/ with a barrel re-export
for backward compatibility.

Modules: common, enums, auth, pagination, route-configs, projects,
labels, variables, stats, pair-groups, characters, world-elements,
gitlab, export-import, user-settings, session.

All 23 importers work unchanged. No circular imports introduced.
